// src/hooks/exchange-rates/types.ts
export interface ExchangeRate {
  id: string;
  currency_code: string;
  forex_buying: number | null;
  forex_selling: number | null;
  banknote_buying: number | null;
  banknote_selling: number | null;
  cross_rate: number | null;
  update_date: string;
}

export interface FetchError extends Error {
  message: string;
}

export type ExchangeRatesMap = Record<string, number>;